Gamestate

Documentation

getTankList :: [(Float, Float, Float, Float, Color4 Float, Int, [Integer])] -> [Tank]

Function to accept tank parameters and call initializeTank in module Tank to initialize tank parameters

getWeaponList :: [(Float, Float, Float, Float, Float, Color4 Float, Color4 Float, Vector3 Float, Float, Float)] -> [WeaponGraphics]

Function to accept weapon parameters and call initializeWeapon in module Weapon to initialize weapon parameters

initializeGamestate :: GameState

Function to give the initial values of the game by the user

getTileMatrix :: [[Tile]]

Function to read tile map parameters from file and return a matrix of tiles

readrow :: [String] -> Int -> [[Tile]] -> [[Tile]]

SubFunction of getTileMatrix to be called for adding a new row in the tilematrix

readcolumn :: [String] -> Int -> [Tile] -> [Tile]

SubFunction of readrow to be called for adding a new column to a row in the tilematrix

tf :: String -> Bool

Function to convert Int To Bool For Tile isObstacle Parameter