Input

Documentation

moveLeft :: Key

To represent the key that makes the tank move left

moveRight :: Key

To represent the key that makes the tank move right

launch :: Key

To represent the key that is responsible for launching a weapon

increasePower :: Key

To represent the key that increases the turret power

decreasePower :: Key

To represent the key that decreases the turret power

increaseAngle :: Key

To represent the key that increases the turret angle

decreaseAngle :: Key

To represent the key that decreases the turret angle

weapon0 :: Key

To represent the key to be pressed for weapon of type 0

weapon1 :: Key

To represent the key to be pressed for weapon of type 1

weapon2 :: Key

To represent the key to be pressed for weapon of type 2