Types

Documentation

data Key

Type to represent different keys on the keyboard

Constructors

Key Key

data Point

Type to represent a point in two dimensions, where each constructor specifies what the point is for

Constructors

Position Float Float

Velocity Float Float

Acceleration Float Float

data Turret

Type for representing the properties of the turrret of a tank

Constructors

Turret

Fields

angle :: Float

The angle of the turret with respect to the tank

power :: Float

The power chosen by the user for launching

data Direction

Type to specify the direction, left or right. Used by weapons and tanks

Constructors

FacingLeft

FacingRight

data TankState

Type to represent the state variables of a tank

Constructors

TankState

Fields

direction :: Direction

The direction the tank is facing

position :: Point

The position of the bottom left corner of the tank

velocity :: Point

The current velocity of the tank

inclineAngle :: Float

The angle of inclination of the tank to the horizontal

turret :: Turret

The state properties of the turret of the tank

data Tank

Type to represent a tank in the game

Constructors

Tank

Fields

tankState :: TankState

The state of the tank

score :: Float

The score of the player using the tank, i.e., the health of the tank

color :: Color4 Float

The colour of the tank, which depends on the player and score

currentWeapon :: Int

The weapon being currently used

weaponCount :: [Integer]

The number of weapons of the current type available

data Tile

Type to represent a tile in the tile map of the game

Constructors

Tile

Fields

tilePosition :: Point

The position of the bottom left corner of the tile on the screen

isObstacle :: Bool

Whether the position of the tile is an obstacle or not

data GameState

Type to represent the state of the game

Constructors

GameState

Fields

tileMatrix :: [[Tile]]

The tile matrix on the screen

tankList :: [Tank]

The list of tanks in the game

weapon :: [WeaponGraphics]

The list of weapons available

chance :: Int

The player whose turn it currently is

noOfPlayers :: Int

The number of players in the game

isAcceptingInput :: Bool

Whether the game is still accepting input, i.e., not over yet

data Weapon

Type to represent the weapons in the game

Constructors

GenericWeapon

Fields

currentPosition :: Point

The current position of the weapon

currentVelocity :: Float

The current velocity of the weapon

launchVelocity :: Float

The velocity at which the weapon was launched

launchAngle :: Float

The angle at which the weapon was launched

velocityMultiplyingFactor :: Float

The multiplying factor to determine the start velocity

currentAngle :: Float

The current angle of the weapon to the horizontal

impactRadius :: Float

The impact radius of the weapon

isLaunched :: Bool

Whether the weapon is currently mid air

hasImpacted :: Bool

Whether the weapon has impacted yet or not

launchDirection :: Direction

The direction of launch of the weapon

data WeaponGraphics

Type to represent the graphics properties of a weapon

Constructors

WeaponGraphics

Fields

weaponPhysics :: Weapon

The state of the weapon

bulletColor :: Color4 Float

The colour of the weapon

turretColor :: Color4 Float

The colour of the turret when this weapon is launched

bulletRotation :: Vector3 Float

Bullets Current Rotation Angle For Display

turretThickness :: Float

The thickness of the turret when this weapon is launched

lengthOfTurret :: Float

The length of the turret when this weapon is launched

tileMatrixRowSize :: Int

Number of rows in the tile matrix

tileMatrixColumnSize :: Int

Number of columns in the tile matrix

widthOfTile :: Float

Width of a tile in the tile matrix

heightOfTile :: Float

Height of a tile in the tile matrix

widthOfTank :: Integer

Width of the tank in terms of the number of tiles

heightOfTank :: Integer

Height of the tank in terms of the number of tiles

powerIncrement :: Float

The amount by which the power changes per key press

angleIncrement :: Float

The amount by which the angle, in radians, changes per key press

isIndexInRange :: [a] -> Int -> Bool

Function to check if the current index is in the range of the tile map or not